+++ database/migrations/2023_05_20_085130_create_operasional_penjualan_sapi_table.php
@@ -6,31 +6,31 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
-    public function up()
-    {
-        Schema::create('operasional_penjualan_sapis', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('id_penjualan_sapi')->unsigned();
-            $table->integer('harga');
-            $table->string('keterangan');
+	/**
+	 * Run the migrations.
+	 *
+	 * @return void
+	 */
+	public function up()
+	{
+		Schema::create('operasional_penjualan_sapis', function (Blueprint $table) {
+			$table->id();
+			$table->foreignId('id_penjualan_sapi')->unsigned();
+			$table->unsignedBigInteger('harga');
+			$table->string('keterangan');
 
-            $table->timestamps();
-            $table->foreign('id_penjualan_sapi')->references('id')->on('penjualan_sapis');
-        });
-    }
+			$table->timestamps();
+			$table->foreign('id_penjualan_sapi')->references('id')->on('penjualan_sapis');
+		});
+	}
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
-    {
-        Schema::dropIfExists('operasional_penjualan_sapi');
-    }
+	/**
+	 * Reverse the migrations.
+	 *
+	 * @return void
+	 */
+	public function down()
+	{
+		Schema::dropIfExists('operasional_penjualan_sapi');
+	}
 };
